+++ test/knapsack/KnapsackTest.java
@@ -9,7 +9,7 @@ public class KnapsackTest {
 	@Test
 	public void test() {
 		int elements[] = {3,5,7};
-		int maxSum = Knapsack.getMaxSum(elements, 3);
+		float maxSum = Knapsack.getMaxSum(elements, 3);
 		System.out.println(maxSum);
 	}
